+++ package/lib/src/querier.dart
@@ -134,11 +134,11 @@ class Querier {
 
           final sample = Sample(
             keyExpr: keyExprStr,
-            payload: utf8.decode(payloadBytes),
+            payload: utf8.decode(payloadBytes, allowMalformed: true),
             payloadBytes: payloadBytes,
             kind: kind == 0 ? SampleKind.put : SampleKind.delete,
             attachment: attachmentBytes != null
-                ? utf8.decode(attachmentBytes)
+                ? utf8.decode(attachmentBytes, allowMalformed: true)
                 : null,
             encoding: encodingStr,
           );
@@ -152,7 +152,7 @@ class Querier {
 
           final replyError = ReplyError(
             payloadBytes: errorPayloadBytes,
-            payload: utf8.decode(errorPayloadBytes),
+            payload: utf8.decode(errorPayloadBytes, allowMalformed: true),
             encoding: errorEncoding,
           );
           controller.add(Reply.error(replyError));