+++ package/lib/src/subscriber.dart
@@ -46,11 +46,11 @@ class Subscriber {
 
         final sample = Sample(
           keyExpr: keyExpr,
-          payload: utf8.decode(payloadBytes),
+          payload: utf8.decode(payloadBytes, allowMalformed: true),
           payloadBytes: payloadBytes,
           kind: kind == 0 ? SampleKind.put : SampleKind.delete,
           attachment: attachmentBytes != null
-              ? utf8.decode(attachmentBytes)
+              ? utf8.decode(attachmentBytes, allowMalformed: true)
               : null,
           encoding: encoding,
         );