+++ package/test/z_ping_shm_cli_test.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:test/test.dart';
 
 /// The direct Dart executable path (not via fvm, to avoid zombie processes).
-const _dartExe = '/home/hugo-bluecorn/fvm/versions/stable/bin/dart';
+final _dartExe = Platform.resolvedExecutable;
 
 /// Forcefully kills a process, using SIGKILL if SIGTERM doesn't work.
 Future<void> forceKill(Process process) async {