+++ package/test/z_queryable_shm_cli_test.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 import 'package:test/test.dart';
 
 /// The FVM-resolved Dart executable path.
-const _dartExe = '/home/hugo-bluecorn/fvm/versions/stable/bin/dart';
+final _dartExe = Platform.resolvedExecutable;
 
 /// Forcefully kills a process, using SIGKILL if SIGTERM doesn't work.
 Future<void> forceKill(Process process) async {