+++ package/test/z_sub_cli_test.dart
@@ -5,7 +5,7 @@ import 'package:test/test.dart';
 import 'package:zenoh/zenoh.dart';
 
 /// The FVM-resolved Dart executable path.
-const _dartExe = '/home/hugo-bluecorn/fvm/versions/stable/bin/dart';
+final _dartExe = Platform.resolvedExecutable;
 
 /// Forcefully kills a process, using SIGKILL if SIGTERM doesn't work.
 Future<void> forceKill(Process process) async {